+++ src/status/compatibleStatusBar.ts
@@ -430,9 +430,9 @@ export class CompatibleStatusBar extends BaseStatusBarItem<CompatibleStatusData>
 
         // 如果距离上次更新不足阈值，则等到满阈值再执行
         const finalDelayMs =
-            timeSinceLastUpdate < this.MIN_DELAYED_UPDATE_INTERVAL
-                ? this.MIN_DELAYED_UPDATE_INTERVAL - timeSinceLastUpdate
-                : delayMs;
+            timeSinceLastUpdate < this.MIN_DELAYED_UPDATE_INTERVAL ?
+                this.MIN_DELAYED_UPDATE_INTERVAL - timeSinceLastUpdate
+            :   delayMs;
 
         StatusLogger.debug(
             `[${this.config.logPrefix}] 设置延时更新提供商 ${providerId}，将在 ${finalDelayMs / 1000} 秒后执行`